Firebase implementation
* Shift from reading data from JSON files to reading them from a Firebase Realtime Database
* Dynamically loaded, so once the database changes app data will change as well without requiring an app reload
* Animated loading icon is displayed while data is being retrieved

NOTE: Code related to shuffling flashcard order was commented out for now because it does not work with firebase.
Data retrieved from firebase is not static, and is being constantly updated, thus, it cannot be manipulated locally.
The code will either have to be removed in the future, or reworked.